requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
bs4==0.0.1
cachetools==5.2.0
catboost==1.0.6
category-encoders==2.5.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
clikit==0.6.2
cloudpickle==2.1.0
cmdstanpy==1.0.4
convertdate==2.4.0
crashtest==0.3.1
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
eif==1.0.2
entrypoints==0.4
ephem==4.1.3
executing==0.10.0
fastjsonschema==2.16.1
flatbuffers==1.12
fonttools==4.35.0
frozenlist==1.3.1
future==0.18.2
gast==0.4.0
gmplot==1.4.1
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.47.0
h5py==3.7.0
hijri-converter==2.2.4
holidays==0.14.2
httpstan==4.8.2
hyperopt==0.2.7
idna==3.3
imageio==2.21.1
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipdb==0.13.9
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.9.1
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
keras==2.9.0
Keras-Preprocessing==1.1.2
keras-vggface==0.6
kiwisolver==1.4.4
korean-lunar-calendar==0.2.1
libclang==14.0.6
lightgbm==3.3.2
llvmlite==0.39.0
LunarCalendar==0.0.9
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.17.0
matplotlib==3.5.3
matplotlib-inline==0.1.3
metrics==0.3.3
mistune==0.8.4
mlmachine==0.1.5
mlxtend==0.20.0
mpmath==1.2.1
multidict==6.0.2
nbclient==0.6.6
nbconvert==6.5.3
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.5
notebook==6.4.12
numba==0.56.0
numexpr==2.8.3
numpy==1.22.4
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pastel==0.2.1
pathlib2==2.3.7.post1
pathspec==0.5.5
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
pkgutil_resolve_name==1.3.10
plotly==5.10.0
prettierplot==0.1.2
prometheus-client==0.14.1
prompt-toolkit==3.0.30
prophet==1.1
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.2.0
pylev==1.4.0
PyMeeus==0.5.11
pyparsing==3.0.9
pyrsistent==0.18.1
pysimdjson==3.2.0
pystan==3.5.0
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
pyzmq==23.2.1
qtconsole==5.3.1
QtPy==2.2.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.0
seaborn==0.11.2
Send2Trash==1.8.0
setuptools-git==1.2
shap==0.41.0
SimpSOM==1.3.4
six==1.16.0
sklearn==0.0
slicer==0.0.7
SOMPY @ git+https://github.com/compmonks/SOMPY.git@5e94ba6d2d5457b3ae231eabed9b63436639d13d
soupsieve==2.3.2.post1
squarify==0.4.3
stack-data==0.4.0
statsmodels==0.13.2
sympy==1.10.1
tenacity==8.0.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.11
wcwidth==0.2.5
webargs==8.2.0
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==3.6.1
wrapt==1.14.1
xgboost==1.6.1
yarg==0.1.9
yarl==1.8.1
zipp==3.8.1
